Enable SSL for mgmt servers and agents
The port remains 8250.

The keystore saved at /etc/cloud/management/cloud.keystore. We also include one
fail-safe keystore/certificate for fallback if we are unable to generate
certificate and keystore. If we use fail-safe keystore, a warning and calltrace would be show.

Notice you need to upgrade agent, as well as systemVM's images.