bug 11305: Stop disconnected redundant virtual router if we can process with the connected one
When we apply rules or start new VM, we may encounter some running routers that
we can't program. That can due to network issue or host is down or vCenter is
disconnected, etc. To keep the synchronization, we would stop them, but only
when there is the other router we've successfully updated. If both routers are
unable to communicate with, we simply give up and report it user.

Conflicts:

	server/src/com/cloud/network/router/VirtualNetworkApplianceManagerImpl.java