bug 14231: Per-account public IP range support
Add UI support for associating IP ranges with specific account + domain;
this is accessible by going to the system network graph->public->IP
range edit, where clicking on 'Add user' button opens a dialog to
specify an account and domain.

Note that right now editing of existing IP ranges is not supported, so
in order to associate a range with another account, the IP range needs
to be dropped and re-created with the new account configuration.