CS-15281: Removal of third party dependencies in Citrix code base.
[Problem]
CloudStack uses a significant amount of third party software.  As part of the move to ASF there is a certain set of licenses that are compatible with ASF policy.  We need to make sure that every dependency we have is in that set. If it's not we have to remove it.

[Solution]
First set: Removing JnetPcap.

[Reviewers]
Edison Su, David Nalley

[Testing]
    [Test Cases]
     Executed ANT build-all sucessfully after removing JnetPcap and its respective dependencies.

    [Platform]
     Fedora release

Signed-off-by: Pradeep <pradeep.soundararajan@citrix.com>