agent: Remove cloud-cleanup.conf for Ubuntu
This is a very dangerous file. Although we do not package it, it would be very dangerous to do so.

If this file would be present on a HyperVisor it would cause all instances to be stopped on a libvirt restart.

There is no need for stopping all instances when libvirt is being stopped (or restarted).