Moved apidocs,marvin and devcloud to developer profile
These maven targets depend on a specific setup of a developer desktop
and require special tools that might not be available on all operating
systems (like bash). Also on of the bash scripts directly calls javac
instead of using the maven build system to deal with dependencies to the
method might not run correcly on all systems.

I've moved these tasks to the developer profile, so you can still use
them but you need to add -P developer to the mvn commandline. This
allows the base of cloudstack to build and run even if you environment
is not configured to run publican for example.