Summary: Sample packaging scripts
This is a manual merge of the maven-to-rpm branch. This commits show how an RPM
package can be built without waf. The current version only builds the management
server rpm and some related rpms like setup. The main missing items are agent rpm and
the awsapi rpm. But it should at least show how to use maven and packaging.

Several small tweaks are put into the client/pom.xml to make sure that the war
has all items required to run as a standalone war.