CLOUDSTACK-1181: Fix Environment so db.properties is in classpath
We export CATALINA_HOME to utils/ which is the path which a lot of utils class would
check and read db.properties file. This is exactly what is done in production.
The utils classes (Transaction, Encryption*Checker etc.) would use PropertiesUtil
which tries to find it in CATALINA_HOME.

The only issue is, even if someone defined an override file, the encryption type
would be still read from the db.properties file. For all other params, the override
file would be re read and datastores would be reinitialized

BUG-ID: CLOUDSTACK-1181

Signed-off-by: Rohit Yadav <bhaisaab@apache.org>