CLOUDSTACK-1347: Don't enforce not null rule on queue_proc_time of sync_item_queue
In c63dbb88042d7eabea2664c2b608c51792fc9f18 I removed the rule from create-schema:
-  `queue_proc_time` datetime COMMENT 'when processing started for the item',

But, upgrade path schema-40to410.sql had a different rule which caused the bug:
+ALTER TABLE `cloud`.`sync_queue_item` ADD `queue_proc_time` DATETIME NOT NULL
COMMENT 'when processing started for the item' AFTER `queue_proc_number`;

In this fix we just revert to whatever rule was defined in create-schema as the
developer may have forgetten to fix same rule in create-schema and upgrade path.
This commit can be reverted or the code be fixed if we want that queue_proc_time
cannot be null.

Signed-off-by: Rohit Yadav <bhaisaab@apache.org>