Install wizard UI: Add preinstall hook
For fresh install, adds hook to install wizard UI to show a custom
'step' before the initial welcome screen.

Pre-install specified by providing 'cloudStack.preInstall,' which
returns a jQuery object to be displayed. It passes args.complete,
which when called will proceed to the first step of the install
wizard.